Cleanup: Separate SSLStats and SSLDiags from SSLUtils
