+++ iocore/net/SSLNetVConnection.cc
@@ -20,22 +20,26 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
+
 #include "tscore/ink_config.h"
 #include "tscore/EventNotify.h"
 #include "tscore/I_Layout.h"
 #include "records/I_RecHttp.h"
+
+#include "InkAPIInternal.h" // Added to include the ssl_hook definitions
+#include "Log.h"
+#include "HttpTunnel.h"
+#include "ProxyProtocol.h"
+#include "HttpConfig.h"
+
 #include "P_Net.h"
 #include "P_SSLNextProtocolSet.h"
 #include "P_SSLUtils.h"
-#include "InkAPIInternal.h" // Added to include the ssl_hook definitions
 #include "P_SSLConfig.h"
-#include "BIO_fastopen.h"
-#include "Log.h"
 #include "P_SSLClientUtils.h"
 #include "P_SSLSNI.h"
-#include "HttpTunnel.h"
-#include "ProxyProtocol.h"
-#include <HttpConfig.h>
+#include "BIO_fastopen.h"
+#include "SSLStats.h"
 
 #include <climits>
 #include <string>
